fix: CameraDetailsPanel, NaverMaps 책임 분리 + 폴더 구조 재정립
